+++ scraper/package.json
@@ -2,19 +2,20 @@
   "name": "ratchagitja-scraper",
   "version": "1.0.0",
   "description": "scraper ratchakitcha from the website",
-  "main": "dist/index.js",
   "scripts": {
     "scrape": "ts-node ./src/index.ts"
   },
-  "author": "",
+  "author": "ntsd",
   "dependencies": {
     "date-fns": "^2.29.3",
     "dayjs": "^1.11.7",
+    "fb": "^2.0.0",
     "node-xlsx": "^0.21.0",
     "object-hash": "^3.0.0",
     "puppeteer": "^19.7.2"
   },
   "devDependencies": {
+    "@types/fb": "^0.0.28",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
   }
